refactor: reorganize sport icons to config directory
Move sport icons configuration from screen utils to global config.
Improves code organization and reusability.

- src/screens/onboardingSportsScreen/utils/sportIcons.ts
  → src/config/sportIcons.ts
